requirements.txt
RPLCD==0.4.0
RPi.GPIO==0.6.3
chardet==2.3.0
colorama==0.3.2
future==0.16.0
html5lib==0.999
oauthlib==2.0.1
python-twitter==3.2.1
requests==2.4.3
requests-oauthlib==0.7.0
six==1.8.0
urllib3==1.9.1
virtualenv==15.1.0
wheel==0.24.0
